+++ src/frontend/src/optimizer/plan_node/logical_agg.rs
@@ -22,10 +22,9 @@ use risingwave_common::error::{ErrorCode, Result, TrackingIssue};
 use risingwave_common::types::DataType;
 use risingwave_common::util::sort_util::OrderType;
 use risingwave_expr::expr::AggKind;
-use risingwave_pb::expr::agg_call::OrderByField as ProstAggOrderByField;
-use risingwave_pb::expr::AggCall as ProstAggCall;
 use risingwave_pb::stream_plan::{agg_call_state, AggCallState as AggCallStateProst};
 
+use super::generic::{PlanAggCall, PlanAggOrderByField, PlanAggCallDisplay};
 use super::{
     BatchHashAgg, BatchSimpleAgg, ColPrunable, LogicalProjectBuilder, PlanBase, PlanRef,
     PlanTreeNodeUnary, PredicatePushdown, StreamGlobalSimpleAgg, StreamHashAgg,
@@ -44,257 +43,6 @@ use crate::optimizer::property::{
 };
 use crate::stream_fragmenter::BuildFragmentGraphState;
 use crate::utils::{ColIndexMapping, Condition, ConditionDisplay, Substitute};
-
-/// Rewritten version of [`crate::expr::OrderByExpr`] which uses `InputRef` instead of `ExprImpl`.
-/// Refer to [`LogicalAggBuilder::try_rewrite_agg_call`] for more details.
-///
-/// TODO(yuchao): replace `PlanAggOrderByField` with enhanced `FieldOrder`
-#[derive(Clone)]
-pub struct PlanAggOrderByField {
-    pub input: InputRef,
-    pub direction: Direction,
-    pub nulls_first: bool,
-}
-
-impl fmt::Debug for PlanAggOrderByField {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{:?}", self.input)?;
-        match self.direction {
-            Direction::Asc => write!(f, " ASC")?,
-            Direction::Desc => write!(f, " DESC")?,
-            _ => {}
-        }
-        write!(
-            f,
-            " NULLS {}",
-            if self.nulls_first { "FIRST" } else { "LAST" }
-        )?;
-        Ok(())
-    }
-}
-
-#[derive(Clone)]
-pub struct PlanAggOrderByFieldDisplay<'a> {
-    pub plan_agg_order_by_field: &'a PlanAggOrderByField,
-    pub input_schema: &'a Schema,
-}
-
-impl fmt::Debug for PlanAggOrderByFieldDisplay<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let that = self.plan_agg_order_by_field;
-        write!(
-            f,
-            "{:?}",
-            InputRefDisplay {
-                input_ref: &that.input,
-                input_schema: self.input_schema
-            }
-        )?;
-        match that.direction {
-            Direction::Asc => write!(f, " ASC")?,
-            Direction::Desc => write!(f, " DESC")?,
-            _ => {}
-        }
-        write!(
-            f,
-            " NULLS {}",
-            if that.nulls_first { "FIRST" } else { "LAST" }
-        )?;
-        Ok(())
-    }
-}
-
-impl PlanAggOrderByField {
-    fn to_protobuf(&self) -> ProstAggOrderByField {
-        ProstAggOrderByField {
-            input: Some(self.input.to_proto()),
-            r#type: Some(self.input.data_type.to_protobuf()),
-            direction: self.direction.to_protobuf() as i32,
-            nulls_first: self.nulls_first,
-        }
-    }
-}
-
-/// Rewritten version of [`AggCall`] which uses `InputRef` instead of `ExprImpl`.
-/// Refer to [`LogicalAggBuilder::try_rewrite_agg_call`] for more details.
-#[derive(Clone)]
-pub struct PlanAggCall {
-    /// Kind of aggregation function
-    pub agg_kind: AggKind,
-
-    /// Data type of the returned column
-    pub return_type: DataType,
-
-    /// Column indexes of input columns.
-    ///
-    /// Its length can be:
-    /// - 0 (`RowCount`)
-    /// - 1 (`Max`, `Min`)
-    /// - 2 (`StringAgg`).
-    ///
-    /// Usually, we mark the first column as the aggregated column.
-    pub inputs: Vec<InputRef>,
-
-    pub distinct: bool,
-    pub order_by_fields: Vec<PlanAggOrderByField>,
-    /// Selective aggregation: only the input rows for which
-    /// `filter` evaluates to `true` will be fed to the aggregate function.
-    pub filter: Condition,
-}
-
-impl fmt::Debug for PlanAggCall {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.agg_kind)?;
-        if !self.inputs.is_empty() {
-            write!(f, "(")?;
-            for (idx, input) in self.inputs.iter().enumerate() {
-                if idx == 0 && self.distinct {
-                    write!(f, "distinct ")?;
-                }
-                write!(f, "{:?}", input)?;
-                if idx != (self.inputs.len() - 1) {
-                    write!(f, ",")?;
-                }
-            }
-            if !self.order_by_fields.is_empty() {
-                let clause_text = self
-                    .order_by_fields
-                    .iter()
-                    .map(|e| format!("{:?}", e))
-                    .join(", ");
-                write!(f, " order_by({})", clause_text)?;
-            }
-            write!(f, ")")?;
-        }
-        if !self.filter.always_true() {
-            write!(
-                f,
-                " filter({:?})",
-                self.filter.as_expr_unless_true().unwrap()
-            )?;
-        }
-        Ok(())
-    }
-}
-
-impl PlanAggCall {
-    pub fn to_protobuf(&self) -> ProstAggCall {
-        ProstAggCall {
-            r#type: self.agg_kind.to_prost().into(),
-            return_type: Some(self.return_type.to_protobuf()),
-            args: self.inputs.iter().map(InputRef::to_agg_arg_proto).collect(),
-            distinct: self.distinct,
-            order_by_fields: self
-                .order_by_fields
-                .iter()
-                .map(PlanAggOrderByField::to_protobuf)
-                .collect(),
-            filter: self
-                .filter
-                .as_expr_unless_true()
-                .map(|expr| expr.to_expr_proto()),
-        }
-    }
-
-    pub fn partial_to_total_agg_call(&self, partial_output_idx: usize) -> PlanAggCall {
-        let total_agg_kind = match &self.agg_kind {
-            AggKind::Min | AggKind::Max | AggKind::StringAgg => self.agg_kind,
-            AggKind::Count | AggKind::Sum | AggKind::ApproxCountDistinct => AggKind::Sum,
-            AggKind::Avg => {
-                panic!("Avg aggregation should have been rewritten to Sum+Count")
-            }
-            AggKind::ArrayAgg => {
-                panic!("2-phase ArrayAgg is not supported yet")
-            }
-        };
-        PlanAggCall {
-            agg_kind: total_agg_kind,
-            inputs: vec![InputRef::new(partial_output_idx, self.return_type.clone())],
-            order_by_fields: vec![], // order must make no difference when we use 2-phase agg
-            filter: Condition::true_cond(),
-            ..self.clone()
-        }
-    }
-
-    pub fn count_star() -> Self {
-        PlanAggCall {
-            agg_kind: AggKind::Count,
-            return_type: DataType::Int64,
-            inputs: vec![],
-            distinct: false,
-            order_by_fields: vec![],
-            filter: Condition::true_cond(),
-        }
-    }
-
-    pub fn with_condition(mut self, filter: Condition) -> Self {
-        self.filter = filter;
-        self
-    }
-
-    pub fn input_indices(&self) -> Vec<usize> {
-        self.inputs.iter().map(|input| input.index()).collect()
-    }
-}
-
-pub struct PlanAggCallDisplay<'a> {
-    pub plan_agg_call: &'a PlanAggCall,
-    pub input_schema: &'a Schema,
-}
-
-impl fmt::Debug for PlanAggCallDisplay<'_> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let that = self.plan_agg_call;
-        write!(f, "{}", that.agg_kind)?;
-        if !that.inputs.is_empty() {
-            write!(f, "(")?;
-            for (idx, input) in that.inputs.iter().enumerate() {
-                if idx == 0 && that.distinct {
-                    write!(f, "distinct ")?;
-                }
-                write!(
-                    f,
-                    "{}",
-                    InputRefDisplay {
-                        input_ref: input,
-                        input_schema: self.input_schema
-                    }
-                )?;
-                if idx != (that.inputs.len() - 1) {
-                    write!(f, ", ")?;
-                }
-            }
-            if !that.order_by_fields.is_empty() {
-                write!(
-                    f,
-                    " order_by({})",
-                    that.order_by_fields.iter().format_with(", ", |e, f| {
-                        f(&format_args!(
-                            "{:?}",
-                            PlanAggOrderByFieldDisplay {
-                                plan_agg_order_by_field: e,
-                                input_schema: self.input_schema,
-                            }
-                        ))
-                    })
-                )?;
-            }
-            write!(f, ")")?;
-        }
-
-        if !that.filter.always_true() {
-            write!(
-                f,
-                " filter({:?})",
-                ConditionDisplay {
-                    condition: &that.filter,
-                    input_schema: self.input_schema,
-                }
-            )?;
-        }
-        Ok(())
-    }
-}
 
 /// `LogicalAgg` groups input data by their group key and computes aggregation functions.
 ///
